+++ invenio_notify/ext.py
@@ -2,7 +2,7 @@ from invenio_notify import config
 from invenio_notify.blueprints import blueprint
 from invenio_notify.resources.config import NotifyInboxResourceConfig
 from invenio_notify.resources.resource import NotifyInboxResource
-from invenio_notify.services.conf import NotifyInboxServiceConfig
+from invenio_notify.services.config import NotifyInboxServiceConfig
 from invenio_notify.services.service import NotifyInboxService
