+++ invenio_notify/services/config.py
